URLSearchParams.prototype.sort():void
As requested in order to fix #22.

The method is patched for native constructors that haven't implemented it yet.
However, even constructoring via records or sequences might be broken.

In those cases there is still no patch or the whole constructor needs to be substituted.

TODO: decide if it's the case to feature detect upfront and globally replace the class if not standard
      also probably return strings like native (quite some work, not sure if worth it)